pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "nativeagents-sdk"
dynamic = ["version"]
description = "Shared contract and primitives for the Native Agents plugin ecosystem"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.11"
authors = [
    { name = "Native Agents", email = "team@nativeagents.net" },
]
keywords = ["claude", "agents", "plugins", "sdk", "ai"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries",
    "Typing :: Typed",
]
dependencies = [
    "pydantic>=2",
    "pyyaml>=6",
    "typer>=0.9",
    "rich>=13",
    "python-dateutil>=2.8",
    "tomli-w>=1.0",
]

[project.urls]
Homepage = "https://github.com/nativeagents/nativeagents-sdk-py"
Repository = "https://github.com/nativeagents/nativeagents-sdk-py"
Issues = "https://github.com/nativeagents/nativeagents-sdk-py/issues"

[project.scripts]
nativeagents-sdk = "nativeagents_sdk.cli.main:app"

[tool.hatch.version]
path = "src/nativeagents_sdk/version.py"

[tool.hatch.build.targets.wheel]
packages = ["src/nativeagents_sdk"]

[tool.pytest.ini_options]
testpaths = ["tests", "tests/e2e"]
addopts = ["-v", "--tb=short", "--cov=nativeagents_sdk", "--cov-branch", "--cov-fail-under=90", "--disable-socket"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "C4", "PIE", "RET", "SIM", "TCH", "ANN", "T20", "FIX002"]
ignore = ["ANN401"]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["ANN", "TC"]

[tool.mypy]
python_version = "3.11"
strict = true
files = ["src/nativeagents_sdk"]
exclude = ["tests/"]

[[tool.mypy.overrides]]
module = ["yaml", "dateutil.*"]
ignore_missing_imports = true

[project.optional-dependencies]
dev = [
    "pytest>=7",
    "pytest-cov>=4",
    "pytest-socket>=0.7",
    "ruff>=0.4",
    "mypy>=1.8",
    "hypothesis>=6",
    "interrogate>=1.5",
    "pip-audit>=2.7",
    "types-PyYAML>=6",
    "types-python-dateutil>=2.8",
]
